Fix fanfare silence on achievement popups (#51) and fullscreen crash with overlays (#52)
- Issue #51: Reset audio stream fade counter (lbl_804D38CC) and restore channel 1 and stream gain in ifprize.c and lbaudio_ax.c so fanfare audio plays at full volume when achievement notices appear
- Issue #51: Immediately update channel volume in synth.c when fade_frames is 0 on TARGET_PC, and add fanfare/notice custom music candidate stems in music_stream.cpp
- Issue #52: Prevent fullscreen launch crashes with NVIDIA overlay / ShadowPlay by guarding window size calculations, aspect ratio divisions, and swapchain reconfiguration against 0x0 and minimized/hidden states in window.cpp and gpu.cpp
- Issue #52: Guard against null texture CreateView dereference and cleanly release surface on Error in aurora.cpp
- Issue #52: Add timeouts and cooperative cancellation checks in WinHTTP updater, and initialize MEM1 prior to launcher/SDL to protect low 32-bit address space

+++ src/pc/disc.h
@@ -71,7 +71,7 @@ static inline void* pc_resolve_dp(uint32_t slot) {
 
 #define DISC_STRUCT __attribute__((scalar_storage_order("big-endian")))
 #define DISC_PTR(T) uint32_t
-#define DP(T, slot) ((T*)pc_resolve_dp((uint32_t)(slot)))
+#define DP(T, slot) ((T*)pc_resolve_dp((uint32_t)(uintptr_t)(slot)))
 #define DP_SET(slot, p)                                                                            \
     do {                                                                                           \
         (slot) = pc_encode_dp((const void*)(p));                                                   \
